requirements.txt
flask>=3.0
gunicorn>=21.0
requests>=2.31
beautifulsoup4>=4.12
schedule>=1.2
python-dotenv>=1.0
lxml>=5.0
